docs(repo-graph): correct ten false claims found by an exhaustive claim audit
Six review rounds had each found exactly one more false assertion in the same
artifacts, so instead of patching a seventh instance this commissioned a
mechanical sweep of every falsifiable claim in the design doc, the commit
messages, the ledger, the touched JSDoc, and the shipping release fragment. Ten
were FALSE and seven OVERSTATED. Corrections, all measured:

- "a type's span is never displaced by a member's" — FALSE, and it SHIPS. The
  rule is positional, not type-preferring: in
  `class X { private void Y() {} private class Y {} }` the key `Y` holds the
  METHOD's span ({2,3}) and the nested type ({4,6}) gets no entry, in Java and
  C# alike. Corrected in the release fragment and in builder.ts.
- "no other language's payload changes" — FALSE, and the shipping fragment
  already said the opposite (correctly). The `exportRanges` widening is gated on
  the grammar id, but a separate fix in the same change added `method_definition`
  to the `javascript` defs query, so `.js` payloads do change.
- "every extracted def" in the exportRanges scope table — FALSE; collisions keep
  one entry and malformed ranges are skipped.
- "None of the three languages permits a raw newline inside an ordinary string or
  char literal" (three sites) — FALSE, compiler-disproved on .NET 10: a C#
  interpolated string spans lines inside an interpolation hole. Narrowed to
  literal TEXT, with the exception stated and why it is harmless (a hole holds
  code, which should not be masked).
- "in all three languages" for the stripComments verbatim defect — FALSE; the
  trigger syntax `@"…"` is C#-only, and the Java and Kotlin analogues resolve
  correctly.
- ontology.ts said a C# raw string "closes on a run of exactly N", contradicting
  both `findClosingQuoteRun`'s own JSDoc and the implementation.
- `sourceBoundaryForLanguage` JSDoc claimed its forms were "verified against the
  real WASM grammars"; the function never loads a grammar — it is an anchored
  regex over masked text.
- `parseCSharpFileImports` JSDoc credited the trailing `;` for excluding C# using
  declarations; the line anchor and dotted-name shape do that work.
- ledger FB-004's rationale ("types.ts is not in the diff" — it is, 23 lines) and
  CO4-04's "Both mutants now die".

That last one was the important one. CO4-04 was itself a finding that branches
survived mutation; its fix introduced a NEW unfalsifiable predicate and then
declared the class closed. Both survivors are now resolved honestly:

- the Java-only escape scoping is PINNED by two new fixtures (a Kotlin and a C#
  raw string with a backslash immediately before the closing fence, where the
  fence must still close because neither language has escapes). Widening the
  predicate now fails 2 tests; removing it fails 1.
- the C#-only verbatim gate is an EQUIVALENT mutant on valid source and is
  documented as such rather than claimed as pinned: `@` must be followed by an
  identifier in Java and Kotlin, so `@"` adjacency in a code position is not
  parseable input. The audit's proposed fixture (`@SuppressWarnings("x")`) has no
  such adjacency and does not discriminate — verified by probe. The gate is kept
  because the masker runs over arbitrary files, including truncated ones.

The remaining "removes that family" overstatement in ontology.ts is replaced by
an explicit list of the three surviving gaps. JLS/JEP citations are marked
UNVERIFIABLE rather than restated as confirmed: no javac or kotlinc is available
on this host, and tree-sitter acceptance is evidence about the grammar, not the
language spec.

The raw-literal tests are split into `repo-graph-raw-literals.test.ts`; the
original file had reached 498 of the 500-line FR-006 cap. No test was dropped.

Merged origin/main (31 commits) so the diff-scoped gates compare against current
main rather than a stale base.
